script: add two config variables
OB_DBMAXSIZE should be set to the default, 536870912, and can be reset
to other sizes. Whatever OB_DBMAXSIZE is set to will show up on
the pvfs2-genconfig command line as the value for filesystem-dbmaxsize.

OB_RUN_IO500... if set to 0 the script will exit after setup and won't
run io500 automatically.

Signed-off-by: Mike Marshall <hubcap@omnibond.com>