Don't add "slots" to hostfile when using Intel MPI
Intel MPI doesn't support specifying the number of slots per host in
the hostfile. So if Intel is being used and a number of slots other
than 0 was specified, print a warning message and instructions on how
to set the number of processes per node. In addition, set IO500 not to
run upon environment creation so as not to run with an incorrect
process mapping.

Signed-off-by: Mike Marshall <hubcap@omnibond.com>