[WIP] Add ability to use digits for instance name suffix
This is the third in a series of commits centered around adding support
for bulk instance-creation.

When we create instances manually, we append a letter to the name of
each instance for differentiation. However, the bulk api automatically
appends numbers to instance names, so we need to be able to choose the
correct suffix so that the correct names are used in the config and
host files and for communicating with instances after creation.