[WIP] Refactor how we handle a few options
This is the fourth in a series of commits centered around adding support
for bulk instance-creation.

Separating the option value from the string that includes the flag name
will allow passing only the option value to the bulk-create script. The
necessity of this will be more apparent with the next commit.