Set the gcloud project property before the region and zone
The property gets updated in the active configuration regardless of
whether the supplied project id is valid or not. With the current
implementation, we set the region and zone before the project. This
leads to unexpected behavior and misleading messages from gcloud when
the script is re-run after having been run with an invalid project.
When run with the invalid project id, we update the property to the
invalid project, gcloud then gives a warning, and we fail later in the
script as expected. However, even if OB_PROJECT is then corrected to a
valid project, the next time the script is run, we try to set the
region and zone first, while the active configuration is still set to
the bad project. Since we are setting Compute Engine properties (region
and zone) on the wrong project id, gcloud gives the following message/
prompt:

API [compute.googleapis.com] not enabled on project [<invalid_project_id>].
Would you like to enable and retry (this will take a few minutes)? (y/N)?

By setting the project property first, we ensure that if OB_PROJECT is
set to a valid project id, then the region and zone will be set with the
correct project configured. And if the script is run with OB_PROJECT set
to an invalid project, then the user will see the above message/prompt
at the appropriate time - when the project id is actually invalid and
immediately after the gcloud warning about the project - better
indicating that the issue is actually with the project id and not the
Compute Engine API.