Fix hostfile issue when using slots with MPICH
MPICH doesn't recognize the OpenMPI format for specifying the number
of slots. If configured for vanilla MPI (MPICH) and a number of slots
is specified, there will be an error parsing the hostfile, causing
IO500 to fail. Fix that by checking the MPI type and crafting the
hostfile entries accordingly.