Don't wait for individual instances when using bulk-create
As of commit 1451114cbf9962919b2e38b147084defb7e9bff2, the bulk_create
script waits for the bulkInsert operation to complete before returning,
so it's no longer necessary to wait for the individual instances in
that case. But we still need to wait until they are running if they
were created one-by-one using basic_create.