Add support for adding custom metadata to instances
By default, each instance is assigned a "type" attribute of either
"server" or "client". This will be useful for startup scripts that need
to perform some actions only on certain instance types. An option for
specifying a startup script will be added in a follow-up commit.
Additional custom metadata attributes can easily be added as needed.