Use strings for metadata instead of associative arrays
Both gcloud and bulk_create.py expect a comma-separated list of metadata
key/value pairs (e.g. "key1=val1,key2=val2"), so it makes more sense to
just start with that rather than comverting an associative array into
the corresponding string. If in the future we find the need to add many
more metadata attributes and/or manipulate individual attributes, then
it might make sense to revisit the arrays, but for now, using strings is
more straightforward.