Add more helper scripts for setting tuning parameters
The recommended way of customizing the settings that these scripts deal
with is via a startup script that is assigned to the instances at
creation; however, these standalone scripts are also provided for cases
when a user wishes to adjust settings after instances have already been
created. They can be run from an instance that has gcloud installed and
has been configured for the gcloud project the instances belong to.