Handle socket timeouts while waiting for an operation to finish
When waiting for larger bulkInsert operations, the default socket
timeout can occur before zoneOperations().wait() returns. In those
cases, the script will prematurely exit in a failed state even if the
operation was successful and the instances were created. Prevent this
by catching the socket.timeout exception and retrying the wait() call.