Add ability to choose which instances use a placement policy
Previously, we only created a single placement policy and applied it to
both client and server instances. This worked when using the
basic_create method of launching instances but not when using
bulk_create. Additionally, when trying to include both servers and
clients in a single policy, especially when using higher numbers of
instances, it isn't always possible to allocate enough resources.
To address those issues and give the user more control over how to use
policies, replace the single shared policy with the option to apply a
single policy to either clients or servers, not both, or to create two
separate policies for each.