Use actual instance names for post-launch setup
When using the basic_create method, it is safe to assume that the first
server or client instance will always use the "a" suffix. If an
instance with that name already exists at run time, creation will fail
and the script will exit. However, that is not the case when using the
bulk_create method. If an instance with the same prefix already exists,
the newly created instances will simply start with the next highest
suffix. So we cannot rely on the client or server instances always
starting at the "01" suffix. We currently make that assumption, so if
an instance with the same prefix already existed, the post-launch setup
will fail because it is trying to perform the setup steps on the old
instance, not the one that was just created. The previous commit added
the functionality to write the actual names of created instances to a
file, so we can solve this issue by reading the names from that file to
perform the setup on the correct instances. Note that since that file
isn't created until instance creation is complete, it is necessary to
move the generation of the OrangeFS configuration file to after that
point, since the config file needs to know the names of the servers.